codeformer-pip/requirements.txt
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.7.1
torchvision
tqdm
yapf
lpips
gdown # supports downloading the large file from Google Drive
fastapi
uvicorn[standard]
piexif
